Fix automatic restarting of the bot on reboot
This was quite a headache and is a bit of a duct tape solution but it seems to work for now.

I'm not very happy with the StartLimitInterval system, I would have wanted it to stop retrying completely if a maximum number of retries is exceeded, but couldn't figure it out.